[common/desktop] time calculations reverted to long, not double

+++ android/src/com/frostwire/android/gui/dialogs/HandpickedTorrentDownloadDialog.java
@@ -244,7 +244,7 @@ public final class HandpickedTorrentDownloadDialog extends AbstractConfirmListDi
         }
 
         @Override
-        public long getItemSize(TorrentFileEntry data) {
+        public double getItemSize(TorrentFileEntry data) {
             return data.getSize();
         }
